backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyJWT==2.8.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-multipart==0.0.6
pydantic==2.5.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0